# unicode/utf8: make accept_range comparison more readable