Hello, the new code has the similar performance and the acceptRange comparison code is easier to read.
```
name                                 old time/op  new time/op  delta
RuneCountTenASCIIChars-4             12.4ns ± 1%  12.4ns ± 1%    ~     (p=0.307 n=8+8)
RuneCountTenJapaneseChars-4          42.6ns ± 1%  42.3ns ± 1%  -0.70%  (p=0.038 n=8+8)
RuneCountInStringTenASCIIChars-4     9.53ns ± 1%  9.54ns ± 1%    ~     (p=0.663 n=8+8)
RuneCountInStringTenJapaneseChars-4  42.4ns ± 1%  42.4ns ± 2%    ~     (p=0.732 n=8+8)
ValidTenASCIIChars-4                 10.4ns ± 0%  10.4ns ± 1%    ~     (p=0.755 n=8+7)
ValidTenJapaneseChars-4              37.1ns ± 2%  37.2ns ± 3%    ~     (p=0.761 n=8+7)
ValidStringTenASCIIChars-4           10.0ns ± 0%  10.1ns ± 1%    ~     (p=0.233 n=6+8)
ValidStringTenJapaneseChars-4        36.9ns ± 2%  36.6ns ± 1%    ~     (p=0.152 n=8+7)
```